Refactored: Logger and Command Line Interface.
Refactored the logger to fit in the "Manager" paradigm introduced for 
all other packages. The CLI now relies on the spf13/cobra package, which 
is much more flexible. Starting the app with the "renderhive -b" command 
line parameters, opens a bash interface where the user can dynamically 
interact with the Renderhive Service App from the command line. Commands 
will be added as the functionality of the app grows.