Added: Official Blender benchmark tool integration.
The official Blender benchmark tool is now supported and can directly be 
launched from the app. At the moment the CLI tool needs to be downloaded 
manually and placed in "benchmark/benchmark-launcher-cli". In the 
future, different versions of this tool will be stored on IPFS/Filecoin 
and will be downloaded by the node when needed. The tool can be called 
from the service app CLI using "node blender benchmark".