Added: Basic mechanism to submit a render request to the render hive.
This is the basic implementation of this function. The corresponding CLI 
command is 'node request submit'. As of now, it sends a message to the 
Renderhive Job Queue topic on the Hedera Consensus Service. Calls to the 
smart contract as well as the implementation of file storage on 
IPFS/Filecoin 
is missing yet.