Added: Calculation of hive cycle start time.
The start timestamp of the current cycle is calculated based on the 
configuration messages in both the consensus time frame of the mirror 
node network and the local time of the Renderhive node. Further 
improvements are required to reduce the number of mirror node queries 
and shift the hive cycle calculation to a more local mechanism.