Added: Sign in and sign out logic.
The previously created basic login page is now working. This enables the basic authentication of the frontend at the backend using the users Hedera wallet. Furthermore, the private key of the node wallet is now encrypted by a passphrase which is derived from a signature that can only be created by the user's wallet. This might later be accompanied by requiring the user to enter a password.