Optimized sign-in logic and added basic sign-up logic
A basic sign-up procedure with Hedera account creation for the new Renderhive node was added. The node account is funded from a user wallet, which is connected to the frontend.  Some functions like smart contract interaction for sign-up and authorization with w3up is still missing. Furthermore, the node data is now encrypted with a user-defined password and the sign-in procedure was adjusted accordigly.