Added: ContractService to the JSON-RPC
This service enables the interaction with smart contracts on Hedera via JSON-RPC calls. Next commits will integrate a JSON-RPC equivalent for each method in the Renderhive Smart Contract. At the moment, this is mainly for testing purposes, but might be partially used in the future for remote control of certain node functions.