Added: Mechanism to prepare Hedera transactions for frontend.
This mechanism enables the backend to prepare any Hedera transaction with the correct parameters without signing or submitting it directly. The transactions bytes can then be send to the frontend for signing by an operator account and submission to the Hedera network, if required. That way, the frontend code can be reduced and all required checks can be offloaded to the backend.